Fix add order (not done)

+++ YC3_DAT_VE_CONCERT/Interface/IPayOSService.cs
@@ -2,7 +2,7 @@
 {
     public interface IPayOSService
     {
-        Task<object> CreatePaymentLink(long orderCode, int amount, string description, string buyerName, string buyerEmail);
+        Task<object> CreatePaymentLink(long orderCode, decimal amount, string description, string buyerName, string buyerEmail);
         Task<object> GetPaymentInfo(long orderCode);
         Task<object> CancelPayment(long orderCode, string? reason = null);
         bool VerifyWebhookSignature(string webhookUrl, string signature);
